requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
django-filter==23.3
drf-yasg==1.21.7
gunicorn==21.2.0

